# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Vigilância Participativa API
  description: API REST para sistema de vigilância participativa em saúde pública
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT no formato Bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalPages:
          type: integer
        totalItems:
          type: integer
    PaginationLinks:
      type: object
      properties:
        first:
          type: string
          format: uri
        last:
          type: string
          format: uri
        prev:
          type: string
          format: uri
          nullable: true
        next:
          type: string
          format: uri
          nullable: true
    User:
      type: object
      required:
        - name
        - email
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    UserCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
          default: true
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
...
    Participation:
      type: object
      required:
        - userId
        - contextId
        - startDate
      properties:
        id:
          type: integer
          readOnly: true
        userId:
          type: integer
        contextId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ParticipationCreate:
      type: object
      required:
        - userId
        - contextId
        - startDate
      properties:
        userId:
          type: integer
        contextId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
          default: true
    ParticipationUpdate:
      type: object
      properties:
        userId:
          type: integer
        contextId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
...
    Report:
      type: object
      required:
        - participationId
        - formVersionId
        - reportType
      properties:
        id:
          type: integer
          readOnly: true
        participationId:
          type: integer
        formVersionId:
          type: integer
        reportType:
          type: string
          enum: [POSITIVE, NEGATIVE]
        description:
          type: string
          nullable: true
        occurrenceLocation:
          type: object
          nullable: true
        responses:
          type: object
          nullable: true
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ReportCreate:
      type: object
      required:
        - participationId
        - formVersionId
        - reportType
      properties:
        participationId:
          type: integer
        formVersionId:
          type: integer
        reportType:
          type: string
          enum: [POSITIVE, NEGATIVE]
        description:
          type: string
          nullable: true
        occurrenceLocation:
          type: object
          nullable: true
        responses:
          type: object
          nullable: true
        active:
          type: boolean
          default: true
    ReportUpdate:
      type: object
      properties:
        participationId:
          type: integer
        formVersionId:
          type: integer
        reportType:
          type: string
          enum: [POSITIVE, NEGATIVE]
        description:
          type: string
          nullable: true
        occurrenceLocation:
          type: object
          nullable: true
        responses:
          type: object
          nullable: true
        active:
          type: boolean
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        links:
          $ref: '#/components/schemas/PaginationLinks'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID do recurso
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Número da página
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamanho da página
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Listar usuários
      description: Retorna lista paginada de usuários
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Users
      summary: Criar usuário
      description: Cria um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Obter usuário por ID
      description: Retorna detalhes de um usuário específico
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Users
      summary: Atualizar usuário
      description: Atualiza um usuário existente
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Users
      summary: Deletar usuário
      description: Remove um usuário
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Usuário deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /participations:
    get:
      tags:
        - Participations
      summary: Listar participações
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: userId
          in: query
          schema:
            type: integer
        - name: contextId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de participações
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Participation'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Participations
      summary: Criar participação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipationCreate'
      responses:
        '201':
          description: Participação criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participation'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /participations/{id}:
    get:
      tags:
        - Participations
      summary: Obter participação
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes da participação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participation'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Participations
      summary: Atualizar participação
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipationUpdate'
      responses:
        '200':
          description: Participação atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participation'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Participations
      summary: Deletar participação
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Participação deletada
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /reports:
    get:
      tags:
        - Reports
      summary: Listar reports
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: participationId
          in: query
          schema:
            type: integer
        - name: formVersionId
          in: query
          schema:
            type: integer
        - name: reportType
          in: query
          schema:
            type: string
            enum: [POSITIVE, NEGATIVE]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Reports
      summary: Criar report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
      responses:
        '201':
          description: Report criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reports/{id}:
    get:
      tags:
        - Reports
      summary: Obter report
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detalhes do report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Reports
      summary: Atualizar report
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportUpdate'
      responses:
        '200':
          description: Report atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Reports
      summary: Deletar report
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Report deletado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'